[WIP] continue work of porting
